+++ frontend/src/pages/RegionalDashboard/index.css
@@ -9,14 +9,6 @@
   }
 }
 
-.ttahub--dashboard-title {
-  font-size: 40px;
-  font-family: 'Merriweather', serif;
-  font-weight: 700;
-  margin: 0 30px 0 0;
-  white-space: inherit;
-}
-
 .ttahub--dashboard-widget-heading {
   font-size: 1.25em;
 }
