+++ frontend/src/pages/ResourcesDashboard/index.scss
@@ -11,14 +11,6 @@
     }
 }
 
-.ttahub--resource-dashboard-title {
-    font-size: 40px;
-    font-family: 'Merriweather', serif;
-    font-weight: 700;
-    margin: 0 30px 0 0;
-    white-space: inherit;
-}
-
 .ttahub--resource-dashboard-widget-heading {
     font-size: 1.25em;
 }
